requirement.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-MQTT==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
paho-mqtt==1.6.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
Werkzeug==3.0.1
